requirements.txt
pdfplumber==0.5.28
pandas==1.3.3
python-docx==0.8.11